pyproject.toml
[project]
name = "vbsocial"
version = "0.1.0"
description = "CLI tool for posting to social media platforms"
authors = [
    {name = "vaibhavblayer", email = "vaibhavblayer@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.0",
    "requests>=2.31.0",
    "requests-oauthlib>=1.3.0",
    "urllib3>=2.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "tqdm>=4.66.0",
]

[project.scripts]
vbsocial = "vbsocial.main:main"

[tool.poetry.scripts]
vbsocial = "vbsocial.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "vbsocial"
version = "0.1.0"
description = "CLI tool for posting to social media platforms"
authors = ["vaibhavblayer <vaibhavblayer@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.12"
click = ">=8.1.0"
requests = ">=2.31.0"
requests-oauthlib = ">=1.3.0"
urllib3 = ">=2.0.0"
google-api-python-client = ">=2.100.0"
google-auth-oauthlib = ">=1.1.0"
tqdm = ">=4.66.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
pytest-cov = ">=4.1.0"
